♻️ [REFACTOR] BattlePage 병합 문제 해결 및 리펙토링 완료
